Move cmd/internal/* to internal/cmd
