+++ internal/cmd/root.go
@@ -1,4 +1,4 @@
-package root
+package cmd
 
 import (
 	"fmt"
